rename: 모달 내 사용되는 option 컴포넌트명 수정 및 import 수정
