+++ app/components/Info/Align/AlignOption.tsx
@@ -1,26 +1,26 @@
-import { filterState } from '@/recoil/Info/atoms';
+import { alignState, filterState } from '@/recoil/Info/atoms';
 import React from 'react';
 import { useRecoilState } from 'recoil';
 import Radio from '@/assets/images/buttons-radio-m.svg';
 import RadioDisabled from '@/assets/images/buttons-radio-m-disable.svg';
 
-const FilterUi = ({
+const AlignOption = ({
   option,
   type,
 }: {
   option: { desc: string; value: string };
-  type: 'petType' | 'signal' | 'signalDepth';
+  type: 'risk' | 'importance';
 }) => {
-  const [filter, setFilter] = useRecoilState(filterState);
+  const [align, setAlign] = useRecoilState(alignState);
   return (
     <div
-      onClick={() => setFilter((prev) => ({ ...prev, [type]: option.value }))}
+      onClick={() => setAlign((prev) => ({ ...prev, [type]: option.value }))}
       className='px-3 py-4 border-b border-extra-deviders flex justify-between'
     >
       <span>{option.desc}</span>
-      {filter[type] === option.value ? <Radio /> : <RadioDisabled />}
+      {align[type] === option.value ? <Radio /> : <RadioDisabled />}
     </div>
   );
 };
 
-export default FilterUi;
+export default AlignOption;
