+++ app/components/Info/Modal/ModalRisk.tsx
@@ -1,26 +1,26 @@
 import Modal, { MODAL_TYPE, MODAL_VARIANT } from '@/components/Modal';
 import React from 'react';
-import { filterList } from '../filterModalList';
-import FilterUi from './FilterUi';
+import { alignList } from '../Filter/filterModalList';
+import AlignOption from '../Align/AlignOption';
 
-const ModalPetType = () => {
-  const modalData = filterList.filter(
-    (item) => item.modalType === MODAL_TYPE.INFO_FILTER_PET_TYPE,
+const ModalRisk = () => {
+  const modalData = alignList.filter(
+    (item) => item.modalType === MODAL_TYPE.INFO_FILTER_RISK,
   )[0];
 
   return (
     <Modal
-      type={MODAL_TYPE.INFO_FILTER_PET_TYPE}
+      type={MODAL_TYPE.INFO_FILTER_RISK}
       variant={MODAL_VARIANT.HALF_SLIDE}
     >
       <Modal.Header title={modalData.title} titleType='left' />
       <div className='px-5 pt-1 pb-5'>
         {modalData.option.map((option) => (
-          <FilterUi key={option.value} type='petType' option={option} />
+          <AlignOption key={option.value} option={option} type='risk' />
         ))}
       </div>
     </Modal>
   );
 };
 
-export default ModalPetType;
+export default ModalRisk;
